requirements.txt
model-security-client
python-dotenv